rename project to 'ethereum'
