+++ setup.cfg
@@ -1,5 +1,5 @@
 [metadata]
-name = eth1spec
+name = ethereum
 description = Ethereum specification, provided as a Python package for tooling
     and testing
 long_description = file: README.md
@@ -12,7 +12,7 @@ classifiers =
     License :: CC0 1.0 Universal (CC0 1.0) Public Domain Dedication
 
 [options]
-packages = eth1spec, eth1spec/evm
+packages = ethereum, ethereum/evm
 package_dir =
     =src
