+++ src/ethereum/evm/ops.py
