+++ src/ethereum/evm/stack.py
