+++ src/ethereum/trie.py
